feat(#16): implement responsive layout for Bento Grid
Closes #16
