+++ lib/src/features/home/presentation/screens/home_screen.dart
@@ -1,15 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:gap/gap.dart';
 import 'package:libri_ai/src/features/home/presentation/widgets/bento_grid.dart';
+import 'package:libri_ai/src/features/home/presentation/widgets/home_app_bar.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
-
-  String _getGreeting() {
-    final hour = DateTime.now().hour;
-    if (hour < 12) return 'Good Morning,';
-    if (hour < 17) return 'Good Afternoon,';
-    return 'Good Evening,';
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -19,88 +14,35 @@ class HomeScreen extends StatelessWidget {
       //   backgroundColor: AppPalette.primary,
       //   child: const Icon(Icons.add, color: Colors.white),
       // ),
-      body: CustomScrollView(
-        slivers: [
-          // 1. Large Collapsing Header
-          SliverAppBar.medium(
-            expandedHeight: 120,
-            floating: false,
-            pinned: true,
-            backgroundColor: const Color(0xFFF8F9FA),
-            surfaceTintColor: Colors.transparent,
-            flexibleSpace: FlexibleSpaceBar(
-              titlePadding: const EdgeInsets.only(left: 20, bottom: 16),
-              centerTitle: false,
-              title: Text.rich(
-                TextSpan(
-                  children: [
-                    TextSpan(
-                      text: '${_getGreeting()}\n',
-                      style: TextStyle(
-                        color: Colors.grey.shade600,
-                        fontSize: 14,
-                        fontWeight: FontWeight.w500,
-                        height: 1.5, // Better line height
-                      ),
-                    ),
-                    const TextSpan(
-                      text:
-                          'Reader', // You could swap this with a Hive username later
-                      style: TextStyle(
-                        color: Colors.black87,
-                        fontSize: 22,
-                        fontWeight: FontWeight.bold,
-                        fontFamily:
-                            'Playfair Display', // Using your Serif font for "Name" looks classy
-                        letterSpacing: -0.5,
-                      ),
-                    ),
-                  ],
-                ),
+      body: LayoutBuilder(
+        builder: (context, constraints) {
+          final width = constraints.maxWidth;
+
+          // 2. Calculate "Tracks" (Cross Axis Count)
+          int crossAxisCount = 4; // Default (Mobile)
+          if (width >= 1000) {
+            crossAxisCount = 8; // Desktop
+          } else if (width >= 600) {
+            crossAxisCount = 6; // Tablet
+          }
+
+          return Center(
+            child: ConstrainedBox(
+              constraints: const BoxConstraints(maxWidth: 1200),
+              child: CustomScrollView(
+                slivers: [
+                  // 1. Large Collapsing Header
+                  const HomeAppBar(),
+
+                  // 2. The Content
+                  HomeBentoGrid(crossAxisCount: crossAxisCount),
+
+                  const SliverGap(20),
+                ],
               ),
             ),
-            actions: [
-              Padding(
-                padding: const EdgeInsets.only(right: 20),
-                child: GestureDetector(
-                  onTap: () {
-                    // Micro-interaction
-                    ScaffoldMessenger.of(context).showSnackBar(
-                      const SnackBar(
-                        content: Text("Profile settings coming soon!"),
-                        duration: Duration(seconds: 1),
-                      ),
-                    );
-                  },
-                  child: Container(
-                    padding: const EdgeInsets.all(2), // White border effect
-                    decoration: BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: Colors.white,
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.1),
-                          blurRadius: 10,
-                          offset: const Offset(0, 4),
-                        ),
-                      ],
-                    ),
-                    child: CircleAvatar(
-                      radius: 22, // Slightly larger
-                      child: Image.asset('images/default_avatar.png'),
-                    ),
-                  ),
-                ),
-              )
-            ],
-          ),
-
-          // 2. The Content
-          const SliverPadding(
-            padding: EdgeInsets.fromLTRB(16, 16, 16, 110),
-            sliver: HomeBentoGrid(),
-          ),
-        ],
+          );
+        },
       ),
     );
   }
